add monitor alerts
